Video Captioning/backend/requirements.txt
flask
torch
transformers
opencv-python
pillow
werkzeug